live-preview: Move all the text edit code into a new module
